feat: bump multiformats to v14 and split identity into ethers-free subpath
- Upgrade multiformats from ^13.4.2 to ^14.0.0, fixing CJS subpath exports
- Extract dependency-free DID functions into src/identity/did-core.ts and expose as @oma3/omatrust/identity/did to address Issue #41

- Also closes Issue #33
