+++ tsup.config.ts
@@ -4,6 +4,7 @@ export default defineConfig({
   entry: [
     "src/index.ts",
     "src/identity/index.ts",
+    "src/identity/did-core.ts",
     "src/reputation/index.ts",
     "src/reputation/index.browser.ts",
     "src/app-registry/index.ts",
